Split the bootique setup depending on whether a database is needed.
